Modificaciones hechas: ** Compatibilidad con Android 2.3. ** Mejorar en el binario del iperf. ** Limpieza de código inservible.

+++ android/app/src/main/java/py/fpuna/tesis/qoetest/utils/VideoUtils.java
@@ -46,7 +46,7 @@ public class VideoUtils {
      * @param totalDuration
      * */
     public int getProgressPercentage(long currentDuration, long totalDuration){
-        Double percentage = (double) 0;
+        Double percentage = 0.0;
 
         long currentSeconds = (int) (currentDuration / 1000);
         long totalSeconds = (int) (totalDuration / 1000);
